requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
google-cloud-firestore==2.13.1
google-cloud-aiplatform==1.38.1
pydantic==2.5.0